+++ src/CollapsibleStack.tsx
